+++ packages/kysely/src/index.ts
@@ -7,7 +7,6 @@ import { createTicketsAdapter } from "./tickets";
 import { createOTPAdapter } from "./otps";
 import { createPasswordAdapter } from "./passwords";
 import { createCodesAdapter } from "./codes";
-import { createUniversalLoginSessionAdapter } from "./universalLoginSessions";
 import { createApplicationsAdapter } from "./applications";
 import { createConnectionsAdapter } from "./connections";
 import { Kysely } from "kysely";
@@ -15,7 +14,6 @@ import { createClientsAdapter } from "./clients";
 import { createKeysAdapter } from "./keys";
 import { createDomainsAdapter } from "./domains";
 import { createBrandingAdapter } from "./branding";
-import { createAuthenticationCodesAdapter } from "./authenticationCodes";
 import { createHooksAdapter } from "./hooks";
 import { createThemesAdapter } from "./themes";
 import { DataAdapters } from "@authhero/adapter-interfaces";
@@ -24,7 +22,6 @@ import { createLoginAdapter } from "./logins";
 export default function createAdapters(db: Kysely<Database>): DataAdapters {
   return {
     applications: createApplicationsAdapter(db),
-    authenticationCodes: createAuthenticationCodesAdapter(db),
     branding: createBrandingAdapter(db),
     clients: createClientsAdapter(db),
     codes: createCodesAdapter(db),
@@ -41,6 +38,5 @@ export default function createAdapters(db: Kysely<Database>): DataAdapters {
     tenants: createTenantsAdapter(db),
     themes: createThemesAdapter(db),
     tickets: createTicketsAdapter(db),
-    universalLoginSessions: createUniversalLoginSessionAdapter(db),
   };
 }